# Cache cast function in set during execution