--- a/src/Interpreters/castColumn.cpp
+++ b/src/Interpreters/castColumn.cpp
@@ -7,24 +7,29 @@ namespace DB
 {
 
 template <CastType cast_type = CastType::nonAccurate>
-static ColumnPtr castColumn(const ColumnWithTypeAndName & arg, const DataTypePtr & type)
+static ColumnPtr castColumn(const ColumnWithTypeAndName & arg, const DataTypePtr & type, InternalCastFunctionCache * cache = nullptr)
 {
     if (arg.type->equals(*type) && cast_type != CastType::accurateOrNull)
         return arg.column;
 
+    const auto & from_name = arg.type->getName();
+    const auto & to_name = type->getName();
     ColumnsWithTypeAndName arguments
     {
         arg,
         {
-            DataTypeString().createColumnConst(arg.column->size(), type->getName()),
+            DataTypeString().createColumnConst(arg.column->size(), to_name),
             std::make_shared<DataTypeString>(),
             ""
         }
     };
+    auto get_cast_func = [&]
+    {
+        FunctionOverloadResolverPtr func_builder_cast = CastInternalOverloadResolver<cast_type>::createImpl();
+        return func_builder_cast->build(arguments);
+    };
 
-    FunctionOverloadResolverPtr func_builder_cast = CastInternalOverloadResolver<cast_type>::createImpl();
-
-    auto func_cast = func_builder_cast->build(arguments);
+    FunctionBasePtr func_cast = cache ? cache->getOrSet(cast_type, from_name, to_name, std::move(get_cast_func)) : get_cast_func();
 
     if constexpr (cast_type == CastType::accurateOrNull)
     {
@@ -36,19 +41,27 @@ static ColumnPtr castColumn(const ColumnWithTypeAndName & arg, const DataTypePtr
     }
 }
 
-ColumnPtr castColumn(const ColumnWithTypeAndName & arg, const DataTypePtr & type)
+ColumnPtr castColumn(const ColumnWithTypeAndName & arg, const DataTypePtr & type, InternalCastFunctionCache * cache)
 {
-    return castColumn<CastType::nonAccurate>(arg, type);
+    return castColumn<CastType::nonAccurate>(arg, type, cache);
 }
 
-ColumnPtr castColumnAccurate(const ColumnWithTypeAndName & arg, const DataTypePtr & type)
+ColumnPtr castColumnAccurate(const ColumnWithTypeAndName & arg, const DataTypePtr & type, InternalCastFunctionCache * cache)
 {
-    return castColumn<CastType::accurate>(arg, type);
+    return castColumn<CastType::accurate>(arg, type, cache);
 }
 
-ColumnPtr castColumnAccurateOrNull(const ColumnWithTypeAndName & arg, const DataTypePtr & type)
+ColumnPtr castColumnAccurateOrNull(const ColumnWithTypeAndName & arg, const DataTypePtr & type, InternalCastFunctionCache * cache)
 {
-    return castColumn<CastType::accurateOrNull>(arg, type);
+    return castColumn<CastType::accurateOrNull>(arg, type, cache);
+}
+
+FunctionBasePtr InternalCastFunctionCache::getOrSet(CastType cast_type, const String & from, const String & to, Getter && getter)
+{
+    std::lock_guard lock{mutex};
+    auto key = std::make_tuple(cast_type, from, to);
+    auto it = impl.find(key);
+    return it != impl.end() ? it->second : impl.emplace(std::forward_as_tuple(cast_type, from, to), getter()).first->second;
 }
 
 }